Add TexView
TexView can take an expression, an equation, a list of them or a tex
code and add the appropriate headers, compile it and show the result.
Appropriate line breaks are inserted so that the lines are not longer
than the textwidth parameter of the actual output. Most things are
configurable.

Also change $TexInitLatexCode to TexInitLatexCode[] which produces the
initial lines of .tex output from $TexInitLatexClassCode, and the lists
$TexInitLatexPackages, $TexInitLatexExtraCode.